Separate git clone commands and it's options from the git command
