+++ Corgibytes.Freshli.Cli/CommandOptions/GitCloneCommandOptions.cs
@@ -1,0 +1,9 @@
+namespace Corgibytes.Freshli.Cli.CommandOptions;
+
+public class GitCloneCommandOptions : CommandOptions
+{
+    public string GitPath { get; set; }
+    public string Branch { get; set; }
+    public string RepoUrl { get; set; }
+}
+
