+++ Corgibytes.Freshli.Cli/Commands/GitCommand.cs
@@ -1,5 +1,4 @@
 using System.CommandLine;
-using Corgibytes.Freshli.Cli.CommandOptions;
 using Corgibytes.Freshli.Cli.Commands.Git;
 
 namespace Corgibytes.Freshli.Cli.Commands;
@@ -14,38 +13,5 @@ public class GitCommand : Command
 
         CheckoutHistoryCommand checkoutHistoryCommand = new();
         AddCommand(checkoutHistoryCommand);
-    }
-}
-
-public class GitCloneCommand : RunnableCommand<GitCloneCommandOptions>
-{
-    public GitCloneCommand()
-        : base("clone", "Clone a repository for Freshli to examine")
-    {
-        Argument<string> repoUrlArgument = new("repo-url", "The URL to the repository to clone")
-        {
-            Arity = ArgumentArity.ExactlyOne
-        };
-
-        AddArgument(repoUrlArgument);
-
-        Option<string> gitPathOption = new("--git-path",
-            description: "The path to the Git executable to use",
-            getDefaultValue: () => "git")
-        {
-            AllowMultipleArgumentsPerToken = false,
-            Arity = ArgumentArity.ExactlyOne,
-        };
-
-        AddOption(gitPathOption);
-
-        Option<string> branch = new("--branch",
-            description: "The branch to check out on the repository")
-        {
-            AllowMultipleArgumentsPerToken = false,
-            Arity = ArgumentArity.ZeroOrOne
-        };
-
-        AddOption(branch);
     }
 }
